Reconfigure header component to be more rigid with content
Previous iteration allowed flexible but difficult to control ChildContent to be rendered in the Content section, and the service name appeared in there rather than next to the logo.
It now follows a similar style to actual GDS pages. If extra flexibility is needed, users can opt not to use the header component.